powerpc/qspinlock: Prepare qspinlock code
I have a bunch of parallel patches that clean up the generic queued
spinlock code, but the powerpc implementation does not use or depend
on any of that except patch collision. I've based the powerpc series
on top of that work, but it's annoying to post or carry around all
those patches as well. This shim patch takes the powerpc changes and
should be applied first. If powerpc series is to go ahead of the
generic series, then this patch would just be merged into patch 1
of this series. This patch won't compile or do anything useful alone.